feat: Add franc language detection to parseLocal function
This commit adds the franc language detection library to the parseLocal function in local.ts. The franc library is used to determine the language of the text extracted from the file. By including language detection, we can provide more accurate metadata about the file, improving the overall functionality of the parsing process.