feat: Add conditional check for NEXT_PUBLIC_ENABLE_RAG in page.tsx
This commit adds a conditional check for the environment variable NEXT_PUBLIC_ENABLE_RAG in the page.tsx file. If the variable is set to 'true', the oracle function is called to retrieve updates and append them to the context variable. This change improves the functionality of the page by enabling the retrieval of additional information when the feature flag is enabled.