refactor: Add Selection component for chat view
This commit adds a new component called Selection to the chat view. The Selection component renders a set of buttons that allow the user to select different options related to enhancing knowledgebase, searching knowledgebase, and analyzing data. This component improves the user experience by providing clear and intuitive options for interacting with the chat interface.